Move rendering of views_savedsearches form to views_post_execute
This was done so we don't render it for input required exposed filters, failed exposed form validation, or empty results. Perhaps we still want to render it for empty results -- a user may wish to save the criteria even if at the moment it produced no results; those criteria may in the future produce results. Perhaps instead of omitting when empty, include as a class so the form may be collapsed or otherwise treated with less prominence.