Avoid sid = 0 when saving
First record written to views_savedsearches table results in a sid of 0. The effects of this were partially addressed in 2c97a34, however it may present problems outside of this module. The flaw exists in views_savedsearches_save when it attempts to calculate the next sid by using "SELECT MAX(sid) + 1"; on an empty table in MySQL, that query returns null. The null value doesn't immediately present a problem because db_query does typecasting on the null value to the %d parameter, which results in a 0 rather than the 1 that was clearly intended by the original query. Moving the increment operation out of the database query allows PHP to type juggle as expected, resulting in the first id off an empty table being 1.

This mearly avoids the problem on new records; it does nothing to fix existing records. No attempt was made to address the problem of existing records to avoid disruption in existing installations and because there may be unknown references to the existing sid. Users can make their own attempts to update if necessary with something like "SELECT MAX(sid) INTO newSid FROM views_savedsearchs; UPDATE views_savedsearches SET sid = newSid WHERE sid = 0; UPDATE x SET sid = newSId WHERE sid = 0;... etc."